ISSUE #5520 - refactor
